Add code to test the MCU
This source code will test the inputs and outputs specified in the
schematic by parsing the number of microsteps informed though UART and
sending the pulses using the timer module. The endstops are also
configured as pull-up inputs with rising edge interruptions enabled and
the stepper motor driver ENABLE' is also driven by the MCU outputs.

This code uses the following resources from the MSP430G2553
microcontroller unit mounted in the LaunchPad MSP-EXP430G2 development
board:

* MCLK and SMCLK at 8 MHz using the internal oscillators.

* P1.0, P1.3, P1.4, P1.5 P1.6 and P1.7 as pull-up enabled inputs with
rising edge interruptions. These are the endstop inputs SW0, SW1, SW2,
SW3, SW4 and SW5.

* P1.1 as UCA0RXD and P1.2 as UCA0TXD. UCA0 is configured as full-duplex
UART module at 9600 bps baud rate. Oversampling (UCOS16 = 1) was used
because the module input clock was high enough to allow the increased
period to evaluate the majority votes, thus reducing the error rate. The
input bytes will be put into a string which will be validated using the
RX interruption from the module.

* P2.0 is set as the output from Timer1_A3.TA0. The pulses will be
generated using the timer module and they will be counted in an
interruption handler routine for the CCR0 chanhel.

* P2.1 is a regular output that will define the rotation direction.

* P2.2, P2.3, P2.4, P2.5 and P2.6 are regular outputs which will drive
the ENABLE' inputs in the stepper motor drivers EN0, EN1, EN2, EN3 and
EN4. Note that EN0 is connected to the two drivers that control the
motors that move the main part of the structure.