Define FSM states
Proposed states for the FSM which will control the system are defined in
fsm_control.h. This Finite State Machine will be executed as soon as the
initial setup is done and will transition states through a function
pointer to be globally available in sys_config.h.

The header file contains the states mentioned and a basic include guard
macro and a basic Doxygen documentation. The functions responsible for
the states are yet to be documented.