Update the available interruption handlers
Remove `step_toggle_ISR()`. The stepper motors now are going to be
driven as regular digital IOs and the timing will be done by polling the
timer flag after the relevant configuration.

The UART interruption service routine will call the command evaluation
function when any of the G-code comments are received (';', '*' or '(').

The `port1_ISR()` was updated to export the endstop that was triggered
and setup the system status flags.