Add extra constants and struct initialization
Define constant for use with the system calibration routine. the speed
will be half as fast as the Z axis stepper motor.

Define constants for the steps per mm data used in the conversion
between absolute mm and 1/32 microsteps.

Declare variable for sending the TX data string.

Add initialization for the system status structs and to leave the vacuum
on, as it is controlled by a normally open valve.

Update Doxygen comments.