Fix axis movements and endstop limitations
The fields `zmax` and `solder_routine` were added to enhance the
debugging experience and to make easer for detecting the safe distance
to the board. There only one endstop in the top of the Z axis, and the
distance to the board will be configured by software.

The calibration routine will only send Z axis to its endstop, as there
is no second endstop and the maximum distances are different and depend
on the system status.

New constants defining the maximum Z distance for solder and regular
operation were defined. The solder flag is set when G0 or G1 is sent
with the E parameter. The maximum z axis distance from the endstop will
define the required position if it is greater than the former.

Movement routine is not limited only by the slow Z axis movement speed.
Now the feedrate will be defined by the slowest motor required in the
movelist.

Update Doxygen documentation.