Make main code run outside the interruptions
Port1 has the lowest priority between the interruptions and would never
be executed if the code is run through interruption handling routines.
In order to fix this problem, the G-code evaluation function is called
by a global variable which is configured after the UART interruption
service routine detects a full buffer or an invallid character.